added mockup gui template
This change consists of adding a new view factory to support the creation of different views, the base controller class for communication between the view and the model, a View abstract class (interface) to define standards for the different views, updated CMake build files, and a wrapper class for threads that makes it easier to enforce threads onto concurrent classes by subclassing the Threads class. This change also adds a resources folder containing a glade file, this file contains the xml for the gui.