Replaced array with map and updated AddFriend
These changes have replaced the fixed sized array with a dynamic map that stores friend nodes indexed by their uuids. This was done to ensure an arbitrary amount of friends can be added, while allowing each one to be checked if it already exists in the map. This provides a more dynamic and scable approach to managing friends.