Updated GetFriend to use map for search
These changes have updated the use of the GetFriend function by replacing the fake return with a request to the map via the given uuid. More so, due to this change the map has been updated to use a shared_ptr. Allowing the GetFriend function to return a shared_ptr. This has been done as it is a safer option to dealing with memory leaks and access control, whilst reducing the memory usage of copying by value.