Updated frontend to work with adding functionality
These changes have reworked how the frontend gui code is layed out to allow for functionality to be easily assigned to buttons dynamically, and linked to the core functionality of the backend. Such as adding friends. More specifically, the observer interface allows for functions to be easily accessed by the guiview and assigned to the buttons provided by the controller as observers. The code has been heavily shifed around as many segmentation faults occured with the previous design. These have now been fixed