Added serialization and deserialization for json
These changes include adding the ability to turn a EVP_PKEY into json, serializing the data ready for transmission along sockets. As well as the ability to take the resulting string and deserialize it back into a EVP_PKEY. This was done to ensure peers in a communication session can easily and efficiently get hold of public keys when needed.