Added missing sections for sending messages
These changes add public key exchange between the server and client, and fix formatting of data to be sent between the server and client. This is achieved by shifting around the json, forcing the send message method to create the json instead of the json being given as a argument.