formatted read for rust server
These changes have modified the read method for server communication to be formatted for communicating with a rust server. This also involves a changing the formatting of the plaintext messages in unit tests and adjusting strings at the end. As well as the way in which the shared key is derived for the unit tests.