create insecure socket state
These changes have added a insecure socket state that implements the socket handler interface. The idea behind this state is to allow for messages to be sent and read without the use of encryption and decryption. This will only ever be used when exchanging the public key for setting up a shared secret.