fix decode of large messages
These changes have fixed a bug where very long messages couldn't be decoded, often leaving huge amounts of data out, causing it to be deleted. As a result the decryption process would always fail. This bug was fixed by changing the use of unsigned char array to a string to ensure the memory is handled in a more efficient manner.