struct to contain data for async read events
This struct is responsible for passing the model, view, and an assoicated connection when a aysnc read event is created. For example, when a socket connection is established, that connection object along with the view and model are given to the event so it may execute behaviour relative to the connection object.