refactor: add virtual desctructor for abstract/interface classes
clang provies compiler warnings for these. a base abstract/interface class needs to be enforced as you can create create an abstract/interface class using a pointer to a derived class, however, when the destructor is invoked it will only call the destructor of that abstract/interface class. This means if the concrete class had some concrete heap allocated memory it won't free as it will only call that super abstract/interface class desctructor and not the dervied one